// src/components/grid-v2/types/grid.ts
export interface ViewportSize {
  width: number
  height: number
}
// ...
export interface UseViewportSizeReturn {
  size: ViewportSize
  containerRef: React.RefObject<HTMLDivElement>
  isInitialized: boolean
}